Split out panel from floaty
